Slight refactoring due to pr comments
